breakout yaml processor to own crate
